codebase-indexer/pyproject.toml
[tool.poetry]
name = "modular-codebase-indexer"
version = "1.0.0"
description = "A modular framework for codebase indexing with pluggable architecture"
authors = ["HealCode Team <team@healcode.com>"]
readme = "README.md"
packages = [{include = "indexer"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
aiofiles = "^23.2.1"
asyncio-mqtt = "^0.16.1"
dependency-injector = "^4.41.0"
pyyaml = "^6.0.1"
watchdog = "^3.0.0"
gitpython = "^3.1.40"
tree-sitter = "^0.20.4"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
networkx = "^3.2.1"
rich = "^13.7.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
coverage = "^7.3.2"

[tool.poetry.scripts]
indexer = "indexer.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 